feat: scan page hardening + AI assistant expansion + docs update
WinUI 3 Scan Page:
- Stop Scan button with process tree cancellation via ScannerService.StopScan()
- Path validation before scan launch with inline error message
- Scan errors display in results panel
- File type distribution chart (FileTypeDistribution model, top 10 extensions)
- Largest files with live filename substring filter
- Export results to JSON from scan page
- Deep/shallow/custom depth scan modes
- Scan speed metrics (files/sec)
- ScannerService process tracking (_currentScannerProcess)

WinUI 3 AI Assistant:
- Expanded tool registry from 3 to 14 tools mirroring Rust backend
- Dynamic ResolveToolChoice() with domain-keyword heuristic
- Enriched ChatRequest with Options, Think, KeepAlive fields

WinUI 3 Converters & Helpers:
- BoolToErrorBrushConverter (red for error states)
- BoolToScanButtonTextConverter (Stop Scan / Start Scan)
- UiHelper.OpenPath() helper
- FileTypeDistribution model

Rust egui scan page:
- Path validation in start_scan()
- Removed unsafe Copy button (no clipboard crate)

Docs: Updated README.md, gui-winui/README.md, AGENTS.md, CHANGELOG.md, docs/README.md

+++ gui-winui/SpaceAnalyzer/Models/DedupModels.cs
@@ -12,7 +12,7 @@ public class DuplicateGroup
 {
     public string Hash { get; set; } = string.Empty;
     public ulong Size { get; set; }
-    public int FileCount { get; set; }
+    public long FileCount { get; set; }
     public List<string> Files { get; set; } = new();
     public ulong WastedBytes { get; set; }
     public string WastedDisplay => ByteFormatter.FormatBytes(WastedBytes);
@@ -25,7 +25,7 @@ public class DuplicateGroup
 public class DedupResult
 {
     public List<DuplicateGroup> DuplicateGroups { get; set; } = new();
-    public int TotalDuplicateFiles { get; set; }
+    public long TotalDuplicateFiles { get; set; }
     public ulong PotentialSavingsBytes { get; set; }
     public string PotentialSavingsDisplay => ByteFormatter.FormatBytes(PotentialSavingsBytes);
 }
